+++ source/Api/Api.csproj
@@ -66,6 +66,7 @@ limitations under the License.
       <ProjectReference Include="..\ArchivedMessages.Application\ArchivedMessages.Application.csproj" />
       <ProjectReference Include="..\BuildingBlocks.Application\BuildingBlocks.Application.csproj" />
       <ProjectReference Include="..\IncomingMessages.Application\IncomingMessages.Application.csproj" />
+      <ProjectReference Include="..\IntegrationEvents\IntegrationEvents.Application\IntegrationEvents.Application.csproj" />
       <ProjectReference Include="..\MasterData.Application\MasterData.Application.csproj"/>
       <ProjectReference Include="..\OutgoingMessages.Application\OutgoingMessages.Application.csproj" />
       <ProjectReference Include="..\Infrastructure\Infrastructure.csproj" />
