+++ source/IntegrationEvents/IntegrationEvents.Application/IntegrationEventHandler.cs
@@ -12,19 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Threading;
-using System.Threading.Tasks;
 using Energinet.DataHub.Core.Messaging.Communication;
 using Energinet.DataHub.EDI.BuildingBlocks.Infrastructure.DataAccess;
-using Energinet.DataHub.EDI.Infrastructure.Configuration.IntegrationEvents;
-using Energinet.DataHub.EDI.Infrastructure.Configuration.IntegrationEvents.IntegrationEventMappers;
+using IntegrationEvents.Infrastructure;
+using IntegrationEvents.Infrastructure.EventProcessors;
 using Microsoft.Extensions.Logging;
 using IIntegrationEventHandler = Energinet.DataHub.Core.Messaging.Communication.Subscriber.IIntegrationEventHandler;
 
-namespace Energinet.DataHub.EDI.Api;
+namespace IntegrationEvents.Application;
 
 #pragma warning disable CA1711
 public sealed class IntegrationEventHandler : IIntegrationEventHandler
