+++ source/IntegrationEvents/IntegrationEvents.Infrastructure/EventProcessors/EnergyResultProducedV2Processor.cs
@@ -12,18 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using Energinet.DataHub.Core.Messaging.Communication;
-using Energinet.DataHub.EDI.Infrastructure.Configuration.IntegrationEvents.IntegrationEventMappers;
 using Energinet.DataHub.EDI.OutgoingMessages.Interfaces;
-using Energinet.DataHub.EDI.Process.Application.Transactions.Aggregations;
 using Energinet.DataHub.Wholesale.Contracts.IntegrationEvents;
 using Microsoft.Extensions.Logging;
 
-namespace Energinet.DataHub.EDI.Process.Application.IntegrationEvents;
+namespace IntegrationEvents.Infrastructure.EventProcessors;
 
 public sealed class EnergyResultProducedV2Processor : IIntegrationEventProcessor
 {
