+++ source/IntegrationEvents/IntegrationEvents.Infrastructure/EventProcessors/MonthlyAmountPerChargeResultProducedV1Processor.cs
@@ -12,22 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Threading;
-using System.Threading.Tasks;
 using BuildingBlocks.Application.FeatureFlag;
 using Energinet.DataHub.Core.Messaging.Communication;
 using Energinet.DataHub.EDI.BuildingBlocks.Infrastructure.DataAccess;
-using Energinet.DataHub.EDI.Infrastructure.Configuration.IntegrationEvents.IntegrationEventMappers;
-using Energinet.DataHub.EDI.OutgoingMessages.Infrastructure.Configuration.DataAccess;
 using Energinet.DataHub.EDI.OutgoingMessages.Interfaces;
-using Energinet.DataHub.EDI.Process.Application.Transactions.Mappers;
-using Energinet.DataHub.EDI.Process.Application.Transactions.WholesaleCalculations;
-using Energinet.DataHub.EDI.Process.Domain.Transactions;
 using Energinet.DataHub.Wholesale.Contracts.IntegrationEvents;
 using Microsoft.EntityFrameworkCore;
 
-namespace Energinet.DataHub.EDI.Process.Application.IntegrationEvents;
+namespace IntegrationEvents.Infrastructure.EventProcessors;
 
 public class MonthlyAmountPerChargeResultProducedV1Processor : IIntegrationEventProcessor
 {
