+++ source/IntegrationEvents/IntegrationEvents.Infrastructure/ReceivedIntegrationEventRepository.cs
@@ -12,14 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
 using System.Data;
-using System.Threading.Tasks;
 using Dapper;
 using Energinet.DataHub.EDI.Common.DateTime;
 using Microsoft.Data.SqlClient;
 
-namespace Energinet.DataHub.EDI.Infrastructure.Configuration.IntegrationEvents;
+namespace IntegrationEvents.Infrastructure;
 
 public class ReceivedIntegrationEventRepository : IReceivedIntegrationEventRepository
 {
