+++ source/IntegrationEvents/IntegrationEvents.Infrastructure/ReceivedIntegrationEventsRetention.cs
@@ -13,8 +13,6 @@
 // limitations under the License.
 
 using System.Data.Common;
-using System.Threading;
-using System.Threading.Tasks;
 using Energinet.DataHub.EDI.BuildingBlocks.Infrastructure;
 using Energinet.DataHub.EDI.BuildingBlocks.Infrastructure.DataAccess;
 using Energinet.DataHub.EDI.Common.DateTime;
@@ -22,7 +20,7 @@ using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.Logging;
 using NodaTime;
 
-namespace Energinet.DataHub.EDI.Infrastructure.Configuration.IntegrationEvents;
+namespace IntegrationEvents.Infrastructure;
 
 public class ReceivedIntegrationEventsRetention : IDataRetention
 {
